Move a bunch of specimen request actions to specimen module
