+++ specimen/src/org/labkey/specimen/view/requestSpecimens.jsp
@@ -28,9 +28,9 @@
 <%@ page import="org.labkey.api.view.JspView" %>
 <%@ page import="org.labkey.api.view.ViewContext" %>
 <%@ page import="org.labkey.api.view.template.ClientDependencies" %>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.CreateSpecimenRequestForm" %>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.HandleCreateSpecimenRequestAction" %>
-<%@ page import="org.labkey.study.controllers.specimen.SpecimenController.NewRequestBean" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController2.CreateSpecimenRequestForm" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController2.HandleCreateSpecimenRequestAction" %>
+<%@ page import="org.labkey.specimen.actions.SpecimenController2.NewRequestBean" %>
 <%@ page import="org.springframework.validation.BindException" %>
 <%@ page import="org.springframework.validation.ObjectError" %>
 <%@ page import="java.util.List" %>
