+++ study/api-src/org/labkey/api/study/StudyInternalService.java
@@ -1,9 +1,13 @@
 package org.labkey.api.study;
 
+import org.labkey.api.action.SpringActionController;
+import org.labkey.api.annotations.Migrate;
 import org.labkey.api.data.Container;
 import org.labkey.api.security.User;
 import org.labkey.api.services.ServiceRegistry;
+import org.labkey.api.specimen.requirements.SpecimenRequest;
 import org.labkey.api.study.model.ParticipantInfo;
+import org.springframework.validation.BindException;
 
 import java.util.Map;
 
@@ -33,4 +37,9 @@ public interface StudyInternalService
     Map<String, ParticipantInfo> getParticipantInfos(Study study, User user, boolean isShiftDates, boolean isAlternateIds);
 
     void generateNeededAlternateParticipantIds(Study study, User user);
+
+    @Migrate
+    void sendNewRequestNotifications(SpringActionController controller, SpecimenRequest request, BindException errors) throws Exception;
+
+    void setLastSpecimenRequest(Study study, Integer lastSpecimenRequest);
 }
