+++ study/src/org/labkey/study/StudyInternalServiceImpl.java
@@ -1,11 +1,16 @@
 package org.labkey.study;
 
+import org.labkey.api.action.SpringActionController;
 import org.labkey.api.data.Container;
 import org.labkey.api.security.User;
+import org.labkey.api.specimen.requirements.SpecimenRequest;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyInternalService;
 import org.labkey.api.study.model.ParticipantInfo;
+import org.labkey.study.controllers.specimen.SpecimenUtils;
+import org.labkey.study.model.StudyImpl;
 import org.labkey.study.model.StudyManager;
+import org.springframework.validation.BindException;
 
 import java.util.Map;
 
@@ -27,5 +32,17 @@ public class StudyInternalServiceImpl implements StudyInternalService
     public void generateNeededAlternateParticipantIds(Study study, User user)
     {
         StudyManager.getInstance().generateNeededAlternateParticipantIds(study, user);
+    }
+
+    @Override
+    public void sendNewRequestNotifications(SpringActionController controller, SpecimenRequest request, BindException errors) throws Exception
+    {
+        new SpecimenUtils(controller).sendNewRequestNotifications(request, errors);
+    }
+
+    @Override
+    public void setLastSpecimenRequest(Study study, Integer lastSpecimenRequest)
+    {
+        ((StudyImpl)study).setLastSpecimenRequest(lastSpecimenRequest);
     }
 }
